+++ spec/unit/provider/reboot/linux_spec.rb
@@ -2,13 +2,26 @@
 require 'spec_helper'
 require 'puppet/type'
 require 'puppet/provider/reboot/linux'
+require 'puppet/provider/reboot/posix'
 
 describe Puppet::Type.type(:reboot).provider(:linux) do
   let(:resource) { Puppet::Type.type(:reboot).new(:provider => :linux, :name => "linux_reboot") }
-  let(:provider) { resource.provider}
+  let(:provider) { resource.provider }
 
   it "should be an instance of Puppet::Type::Reboot::ProviderLinux" do
     provider.must be_an_instance_of Puppet::Type::Reboot::ProviderLinux
+  end
+
+  it "should be a kind of Puppet::Type::Reboot::ProviderPosix" do
+    provider.must be_a_kind_of Puppet::Type::Reboot::ProviderPosix
+  end
+
+  context '#initialize' do
+    it 'should issue a deprecation warning' do
+      Puppet.expects(:deprecation_warning).with("The 'linux' reboot provider is deprecated and will be removed; use 'posix' instead.")
+
+      Puppet::Type.type(:reboot).new(:provider => :linux, :name => "linux_reboot")
+    end
   end
 
   context "self.instances" do
@@ -51,13 +64,24 @@ describe Puppet::Type.type(:reboot).provider(:linux) do
       provider.reboot
     end
 
+    before :each do
+      Facter.stubs(:value).with(:kernel).returns('Linux')
+    end
+
     it "includes the restart flag" do
       provider.expects(:async_shutdown).with(includes('-r'))
       provider.reboot
     end
 
     it "includes a timeout in the future" do
-      provider.expects(:async_shutdown).with(includes("+#{resource[:timeout].to_i / 60}"))
+      provider.expects(:async_shutdown).with(includes("+#{(resource[:timeout].to_i / 60.0).ceil}"))
+      provider.reboot
+    end
+
+    it "rounds up and provides a warning if the timeout is not a multiple of 60" do
+      resource[:timeout] = 61
+      Puppet.expects(:warning).with(includes('rounding'))
+      provider.expects(:async_shutdown).with(includes("+#{(resource[:timeout].to_i / 60.0).ceil}"))
       provider.reboot
     end